Fully Implement Tuplets (#33)
* Creating Design Doc for Tuplets

A basic implementation of Tuplet was done, but there were many cases not covered. This design document will go over all the cases that need to be covered and how the API should work.

* Added stored properties to API

I don't like the naming for the stored properties, but this is what I have so far. If anyone has suggestions, that would be great.

* Changed names to avoid conflict

Names conflicted with the `NoteCollection` protocol, so I changed them slightly.

* Adding Compound Tuplets

Just found out that you can have a compound tuplet. This means you can have a tuplet made up of another tuplet. Added a section in the design document to address this.

* Update tuplets.md

Updated after talking to @migue48

* Spelling fix

* Add replaceNote for Tuplet

Added method replacing a note with a tuplet for a compound tuplet. Then, added some detail about the implementation needed for that.

* Adding link that is not yet active

Added link to indexing-methodology design document, which has not been pushed yet. However, the link should work once it is.

* Update design doc for NoteDuration changes

Documenting changes to `NoteDuration` and why they were made. The changes to `NoteDuration` happened in a separate commit on a separate PR.

* Added standard ratios, modified API

Also added more info about indexes.

* Fixed english errors

* Started tuplet changes. WIP. Not building yet.

* Added standard ratios

* Implemented init tests for Tuplet. Implemented recomputeFlatIndexes.

* Successfully building

* Implemented init validation. Tuplet tests pass

* Implemented replaceNote(at index, with: Note) and its tests